improve contacts permission handling
