+++ FlowCrypt/Controllers/Compose/Extensions/ComposeViewController+Contacts.swift
@@ -18,6 +18,7 @@ extension ComposeViewController {
                     for: .gmailLogin(self),
                     appContext: appContext
                 )
+                contactsProvider.authorization = googleAuthManager.authorization
                 shouldEvaluateRecipientInput = true
                 reload(sections: [.contacts])
             } catch {
