+++ FlowCrypt/Functionality/Mail Provider/Contacts Provider/ContactsProviderType.swift
@@ -7,7 +7,10 @@
 //
 
 import Foundation
+import GTMAppAuth
 
 protocol ContactsProviderType {
+    var authorization: GTMAppAuthFetcherAuthorization? { get set }
+    var isContactsScopeEnabled: Bool { get }
     func searchContacts(query: String) async throws -> [Recipient]
 }
